Feature/firebase analytics

+++ lib/features/body_check/view_models/body_check_result_view_model.dart
@@ -1,6 +1,8 @@
+import 'package:extended_image/extended_image.dart' as extended_image;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:thunder/core/services/analytics_service.dart';
 import 'package:thunder/features/body_check/models/data/body_check_result.dart';
-import 'package:thunder/features/body_check/repsitories/body_check_repository.dart';
+import 'package:thunder/features/body_check/repositories/body_check_repository.dart';
 
 class BodyCheckResultViewModel
     extends AutoDisposeFamilyAsyncNotifier<BodyCheckResult, int> {
@@ -17,6 +19,7 @@ class BodyCheckResultViewModel
   Future<BodyCheckResult> _fetchBodyCheckResult(int bodyPhotoId) async {
     try {
       final data = await _repository.getBodyCheckResult(bodyPhotoId);
+      AnalyticsService.viewBodyResult(bodyPhotoId);
       return BodyCheckResult.fromJson(data);
     } catch (e) {
       state = AsyncError(e, StackTrace.current);
@@ -31,10 +34,16 @@ class BodyCheckResultViewModel
     state = AsyncData(await _fetchBodyCheckResult(bodyPhotoId));
   }
 
-  Future<bool> deleteBodyCheckResult(int bodyPhotoId) async {
+  Future<bool> deleteBodyCheckResult() async {
+    final bodyPhotoId = state.valueOrNull?.bodyPhotoId;
+    final url = state.valueOrNull?.imageUrl;
+    if (bodyPhotoId == null || url == null) return false;
     state = const AsyncLoading();
     try {
       await _repository.deleteBodyCheckResult(bodyPhotoId);
+      AnalyticsService.deleteContent(bodyPhotoId);
+      await extended_image.clearDiskCachedImage(url);
+      extended_image.clearMemoryImageCache(url);
       return true;
     } catch (e) {
       state = AsyncError(e, StackTrace.current);
